Reorganize src/ into core/ engine and apps/ contexts (issue #19)
Move view components and types into src/core/, extract ResearchApp
from App.tsx into src/apps/research/, relocate data files, and add
src/apps/blog/BlogApp.tsx stub. App.tsx is now a thin router shell.
