+++ src/apps/blog/BlogApp.tsx
@@ -1,0 +1,5 @@
+export default function BlogApp() {
+  return (
+    <div className="placeholder">Blog reader — coming soon</div>
+  );
+}
