+++ src/apps/research/data/schema.md
